complete clippy.js
Initially, I wanted to make it a real assistant for users, but I had difficulties connecting a real AI to the page via the api. Then I came across an article where it was said that Skippy the cat had a vulnerability that allowed an attacker to gain control of the system, add and delete files. this was the idea for him to become the fifth computer problem. and yes, do not forget that he already infuriated everyone :) Even I caught him in the pc